+++ src/MP.Server/ApiServer.BLL/IBLL/ISysUserService.cs
@@ -9,11 +9,11 @@ namespace ApiServer.BLL.IBLL
         MsgModel GetUserByUserName(string userName);
         MsgModel UpdateUser(sys_user sys_user);
         MsgModel AddUser(sys_user sys_user);
-        MsgModel DeleteUser(long userId);
-        MsgModel PwdReset(long userId);
+        MsgModel DeleteUser(int userId);
+        MsgModel PwdReset(int userId);
         MsgModel IsDefault(string userName);
         MsgModel ChangePwd(string userName, string oldPass, string newPass);
-        MsgModel UpdateEnabled(long id, bool enabled);
-        MsgModel QueryUser(long? orgId, string userName, string phone, string email, bool? enabled, DateTime? createStartTime, DateTime? createEndTime, int pageNum, int pageSize);
+        MsgModel UpdateEnabled(int id, bool enabled);
+        MsgModel QueryUser(int? orgId, string userName, string phone, string email, bool? enabled, DateTime? createStartTime, DateTime? createEndTime, int pageNum, int pageSize);
     }
 }
