+++ src/MP.Server/ApiServer.Model/Entity/sys_menu.cs
@@ -9,8 +9,8 @@ namespace ApiServer.Model.Entity
 {
     public partial class sys_menu
     {
-        public long id { get; set; }
-        public long menu_pid { get; set; }
+        public int id { get; set; }
+        public int menu_pid { get; set; }
         public string menu_pids { get; set; }
         public bool is_leaf { get; set; }
         public string menu_name { get; set; }
