+++ src/MP.Server/ApiServer.Model/Entity/sys_org.cs
@@ -9,8 +9,8 @@ namespace ApiServer.Model.Entity
 {
     public partial class sys_org
     {
-        public long id { get; set; }
-        public long org_pid { get; set; }
+        public int id { get; set; }
+        public int org_pid { get; set; }
         public string org_pids { get; set; }
         public bool is_leaf { get; set; }
         public string org_name { get; set; }
