+++ src/MP.Server/ApiServer.Model/Entity/sys_role.cs
@@ -9,7 +9,7 @@ namespace ApiServer.Model.Entity
 {
     public partial class sys_role
     {
-        public long id { get; set; }
+        public int id { get; set; }
         public string role_name { get; set; }
         public string role_desc { get; set; }
         public string role_code { get; set; }
