+++ src/MP.Server/ApiServer.Model/Entity/sys_user_role.cs
@@ -9,8 +9,8 @@ namespace ApiServer.Model.Entity
 {
     public partial class sys_user_role
     {
-        public long id { get; set; }
-        public long role_id { get; set; }
-        public long user_id { get; set; }
+        public int id { get; set; }
+        public int role_id { get; set; }
+        public int user_id { get; set; }
     }
 }
